seperate components to different files
